Put QueryTools and ActionTools in distinct model file
