+++ projects/ngx-testing-extra/src/lib/components/extra-bed/models/action-tools.model.ts
@@ -1,0 +1,9 @@
+import { Type } from '@angular/core';
+
+export interface ActionTools {
+  click(selector: string): void;
+
+  click(directive: Type<any>): void;
+
+  emitChildOutput(directive: Type<any>, outputName: string, outputValue?: any): void;
+}
