+++ projects/ngx-testing-extra/src/lib/components/extra-bed/models/index.ts
@@ -1,1 +1,1 @@
-export { ExtraTools, ExtraOptions } from './extra.models';
+export { ExtraTools, ExtraOptions } from './extra-bed.models';
